+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.3] - 2026-06-13
 
 ### Added
 
@@ -252,6 +252,7 @@ Initial alpha release.
 - Renamed `Authentication` class to `AuthenticationLogin`
 - Renamed `clear` method to `failedAuthentication` in Token
 
+[2.0.3]: https://github.com/oxygensuite/oxygen-ergani/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/oxygensuite/oxygen-ergani/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/oxygensuite/oxygen-ergani/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/oxygensuite/oxygen-ergani/compare/v1.1.1...v2.0.0